backend/pyproject.toml
[project]
name = "crewai_financial_analyst"
version = "0.1.0"
description = "crewai_financial_analyst using crewAI"
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai[tools]>=0.76.9,<1.0.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=1.0.0",
    "google-search-results>=2.4.2",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "openai>=1.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
